splitting elasticsearch and simple search code
