+++ backend/tracim_backend/views/search_api/simple_search_controller.py
@@ -1,0 +1,5 @@
+from tracim_backend.views.search_api.search_controller import SearchController
+
+
+class SimpleSearchController(SearchController):
+    pass
